Relocation of plot handler and updates to unit tests
